requirements.txt
PyMuPDF
tiktoken
numpy
openai
pinecone
sentence-transformers
gradio
torch
python-dotenv
matplotlib